Support interactive mode for exec command
The support of interactive execute is similar to attach but
the docker exec endpoint doesn't support websocket as the attach
endpoint. Therefore, this commit introduced an HTTP client for
streaming. The common code between websocketclient and HTTP
client are consolidated into a base class.

This patch relies on the server to return the URL of Docker
Engine [1] and leverage the URL to connect to the container and
execute commands interactively. A sample workflow would be as
following.

1. Create a container that never exited (by using the command
   "tail -f /dev/null"):
   [host]$ zun run --name test ubuntu tail -f /dev/null

2. Enter the container with a Bash shell:
   [host]$ zun exec -i test bash

3. Once you are inside, you can run any command you want:
   [container]# apt-get update

4. Once finished, exit the container:
   [container]# exit

[1] https://review.openstack.org/#/c/452505/

Implements: blueprint support-interactive-mode
Depends-On: If987222ecf5419a305810858e304a4c2b4a78474
Change-Id: If02b20df46496d2687e60fefce46046d7377f441