Fix confusing error message on interactive run
If run a container with an incorrect command in interactive mode,
it showed a confusing error message "ERROR: Invalid websocket link when attach ...".
This error message is sub-optimal because the root cause of the failure is the container
entered an error state (because users specify an incorrect command).
We need to fix it by checking the state of the container and failed it
with a correct message if the container entered an error state.

Change-Id: I630de2414d5e36e86b246460a6d24992b06696e3
Closes-Bug: #1685521