Upgrade from docker-py to docker
The pypi package 'docker-py' [1] has been renamed to 'docker' [2].
We need to move to the new 'docker' package in order to leverage
the up-to-date features.

NOTE: In order to pass the gate, the OSC tests were temporarily
disabled. Those tess will be re-enabled once both server and
client finish the upgrade.

[1] https://pypi.python.org/pypi/docker-py
[2] https://pypi.python.org/pypi/docker
[3] https://review.openstack.org/#/c/423715/

Change-Id: I9346112d8a495bc6b304647377f8da47017cc869
Partial-Implements: blueprint adopt-docker-py-2.0-api