Generate stestr.conf
When run functional tests, a warning was raised:

UserWarning: No .stestr.conf file found in the CWD. Please create one
to to replace the .testr.conf.

This patch creates a .stestr.conf file but keep .testr.conf. Will
remove it & its usage when testrepository is completely replaced by
stestr. OpenStack has a long-term plan for this [1].

[1] https://review.openstack.org/#/c/491075/

Change-Id: I3507c381551d153fc6ca742b8bd4b8333710580b