Wait for container deletion in tests
Container deletion was changed from sync to async. If users submit a
request to delete a container, server will response immediately but
the container might not be deleted right away. Users should not assume
the container is deleted unless the container is not shown up in list.

Therefore, the tempest test case need to be updated to deal with this
change. After the delete request is issued, we add logic to wait for
the container to disappear before doing the assertion. This should
fix the intermittent failure of the gate.

Change-Id: I9160c503183a1d909dfcfe67eba0e27ae3499a14