Change host list display value
This patch add disk_total in host list display,
and delete os and labels in host list display.
Whether the host is available can be determined
by CPU, memory, and disk. OS and labels not high
in the value available to the host.
Use "host-show" get host detail message, The os and
labels will display.

Change-Id: I6bd47ebb84855485c60bf079cc6cd6efe483c814
Signed-off-by: Yuanbin.Chen <cybing4@gmail.com>