Re-introduce gate_hook.sh in zunclient
The script gate_hook.sh was removed at Zun side [1] as part of
the effort for migrating to zuul v3 native job. However, zunclient
still need this script and the removal causes the failure of the
CI. This patch re-introduce the script at zunclient as a temporary
fix. The long-term fix is to migrate the job to zuul v3 native
as well.

[1] https://review.openstack.org/#/c/512209/

Change-Id: I93a4ec8804e74e17c478f2ee2ecf9b3402fcbd3f