Encode injected file data in containers module
We encode file content before sending it via HTTP. In before,
we encode injected file content in CLI shell module so the file
is encoded only when users are using CLI commands.
This commit moves encoding down to containers module so that it
will be available for non-CLI users as well.

Change-Id: I9638ee363ae85c36e421dc997fe2dcad5d4d71cc